pyproject.toml
[tool.poetry]
name = "Samdash"
version = "0.1.0"
description = "My personal dashboard."
authors = ["Samuel Morris <dodobird181@gmail.com>"]
license = "The Unlicense"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12.3"
django = "^5.0"
djangorestframework = "^3.15"
psycopg2-binary = "^2.9"
feedparser = "^6.0"
requests = "^2.32"
python-decouple = "^3.8"
whitenoise = "^6.7"
django-cors-headers = "^4.4"
google-auth = "^2.35"
gunicorn = "^23.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pytest-django = "^4.9"
coverage = "^7.6"
factory-boy = "^3.3"
faker = "^33.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.test_settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
